dedup sample_crop_data
