+++ tests/test_feed_storage_module/test_feed_manager.py
@@ -3,49 +3,12 @@ import pytest
 from RUFAS.routines.feed_storage.feed_manager import FeedManager, StorageType
 from RUFAS.routines.feed_storage.harvested_crop import HarvestedCrop
 from RUFAS.routines.feed_storage.enums import CropCategory, CropType
-from RUFAS.time import Time
-from RUFAS.config import Config
+
+from .sample_crop_data import sample_crop_data  # noqa F401
 
 
 @pytest.fixture
-def sample_crop_data() -> Dict[str, float]:
-    return {
-        "harvest_time": Time(
-            Config(
-                {
-                    "start_date": "1:1",
-                    "end_date": "1:10",
-                    "set_seed": False,
-                    "random_seed": 42,
-                }
-            )
-        ),
-        "storage_time": Time(
-            Config(
-                {
-                    "start_date": "1:1",
-                    "end_date": "1:10",
-                    "set_seed": False,
-                    "random_seed": 42,
-                }
-            )
-        ),
-        "fresh_mass": 100.0,
-        "dry_matter_percentage": 50.0,
-        "dry_matter_digestibility": 70.0,
-        "crude_protein_percent": 10.0,
-        "non_protein_nitrogen": 5.0,
-        "starch": 30.0,
-        "adf": 7.0,
-        "ndf": 15.0,
-        "lignin": 3.0,
-        "sugar": 20.0,
-        "ash": 6.0,
-    }
-
-
-@pytest.fixture
-def harvested_crop(sample_crop_data: Dict[str, float]) -> HarvestedCrop:
+def harvested_crop(sample_crop_data: Dict[str, float]) -> HarvestedCrop:  # noqa F811
     """
     Pytest fixture to create a HarvestedCrop instance for testing.
 
@@ -93,8 +56,8 @@ def test_receive_crop_multiple(
         )
         assert StorageType.DRY in feed_manager.active_storages.keys()
         assert StorageType.BUNKER in feed_manager.active_storages.keys()
-        assert len(feed_manager.active_storages[StorageType.BUNKER].stored) ==1
-        len(feed_manager.active_storages[StorageType.DRY].stored)==2
+        assert len(feed_manager.active_storages[StorageType.BUNKER].stored) == 1
+        len(feed_manager.active_storages[StorageType.DRY].stored) == 2
         pass
     except ValueError:
         pytest.fail("Unexpected ValueError raised")
