+++ tests/test_feed_storage_module/test_harvested_crop.py
@@ -2,45 +2,7 @@ from typing import Dict
 import pytest
 from RUFAS.routines.feed_storage.harvested_crop import HarvestedCrop
 from RUFAS.routines.feed_storage.enums import CropCategory, CropType
-from RUFAS.time import Time
-from RUFAS.config import Config
-
-
-@pytest.fixture
-def sample_crop_data() -> Dict[str, float]:
-    return {
-        "harvest_time": Time(
-            Config(
-                {
-                    "start_date": "1:1",
-                    "end_date": "1:10",
-                    "set_seed": False,
-                    "random_seed": 42,
-                }
-            )
-        ),
-        "storage_time": Time(
-            Config(
-                {
-                    "start_date": "1:1",
-                    "end_date": "1:10",
-                    "set_seed": False,
-                    "random_seed": 42,
-                }
-            )
-        ),
-        "fresh_mass": 100.0,
-        "dry_matter_percentage": 50.0,
-        "dry_matter_digestibility": 70.0,
-        "crude_protein_percent": 10.0,
-        "non_protein_nitrogen": 5.0,
-        "starch": 30.0,
-        "adf": 7.0,
-        "ndf": 15.0,
-        "lignin": 3.0,
-        "sugar": 20.0,
-        "ash": 6.0,
-    }
+from .sample_crop_data import sample_crop_data  # noqa F401
 
 
 @pytest.mark.parametrize(
@@ -65,7 +27,9 @@ def sample_crop_data() -> Dict[str, float]:
     ],
 )
 def test_valid_category_type_combinations(
-    category: CropCategory, crop_type: CropType, sample_crop_data: Dict[str, float]
+    category: CropCategory,
+    crop_type: CropType,
+    sample_crop_data: Dict[str, float],  # noqa F811
 ):
     try:
         HarvestedCrop(category=category, type=crop_type, **sample_crop_data)
@@ -83,13 +47,15 @@ def test_valid_category_type_combinations(
     ],
 )
 def test_invalid_category_type_combinations(
-    category: CropCategory, crop_type: CropType, sample_crop_data: Dict[str, float]
+    category: CropCategory,
+    crop_type: CropType,
+    sample_crop_data: Dict[str, float],  # noqa F811
 ):
     with pytest.raises(ValueError):
         HarvestedCrop(category=category, type=crop_type, **sample_crop_data)
 
 
-def test_attributes(sample_crop_data: Dict[str, float]):
+def test_attributes(sample_crop_data: Dict[str, float]):  # noqa F811
     crop = HarvestedCrop(
         category=CropCategory.SMALL_GRAIN, type=CropType.WHEAT, **sample_crop_data
     )
